group components, re-implement with stateful classes
